Creates LFG thread in forum channel
Implements the creation of a new "Looking For Group" (LFG) thread within a designated forum channel.

The thread includes an embed with details such as gameplay mode, region, team description, language, and time.
It also applies appropriate tags to the thread and updates the user's modal reply with a link to the newly created thread.